pyproject.toml
[tool.poetry]
name = "ansaction-analysis"
version = "0.1.0"
description = ""
authors = ["Denis Guryanov <guryanov19962010@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"


[tool.poetry.group.tests.dependencies]
pytest = "^8.3.4"


[tool.poetry.group.test.dependencies]
freezegun = "^1.5.1"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'